ai_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.26.4
requests==2.31.0
pandas==2.0.3
